Fix navigation_bar examples
