+++ examples/api/lib/material/platform_menu_bar/platform_menu_bar.0.dart
@@ -11,32 +11,32 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-void main() => runApp(const SampleApp());
+void main() => runApp(const ExampleApp());
 
 enum MenuSelection {
   about,
   showMessage,
 }
 
-class SampleApp extends StatelessWidget {
-  const SampleApp({super.key});
+class ExampleApp extends StatelessWidget {
+  const ExampleApp({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const MaterialApp(
-      home: Scaffold(body: MyMenuBarApp()),
+      home: Scaffold(body: PlatformMenuBarExample()),
     );
   }
 }
 
-class MyMenuBarApp extends StatefulWidget {
-  const MyMenuBarApp({super.key});
+class PlatformMenuBarExample extends StatefulWidget {
+  const PlatformMenuBarExample({super.key});
 
   @override
-  State<MyMenuBarApp> createState() => _MyMenuBarAppState();
+  State<PlatformMenuBarExample> createState() => _PlatformMenuBarExampleState();
 }
 
-class _MyMenuBarAppState extends State<MyMenuBarApp> {
+class _PlatformMenuBarExampleState extends State<PlatformMenuBarExample> {
   String _message = 'Hello';
   bool _showMessage = false;
