+++ examples/api/test/material/navigation_bar/navigation_bar.0_test.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:flutter/material.dart';
-import 'package:flutter_api_samples/material/navigation_bar/navigation_bar.dart'
+import 'package:flutter_api_samples/material/navigation_bar/navigation_bar.0.dart'
     as example;
 import 'package:flutter_test/flutter_test.dart';
 
@@ -11,7 +11,7 @@ void main() {
   testWidgets('Navigation bar updates destination on tap',
       (WidgetTester tester) async {
     await tester.pumpWidget(
-      const example.MyApp(),
+      const example.ExampleApp(),
     );
     final NavigationBar navigationBarWidget =
         tester.firstWidget(find.byType(NavigationBar));
