+++ examples/api/test/material/navigation_bar/navigation_bar.1_test.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:flutter/material.dart';
-import 'package:flutter_api_samples/material/navigation_bar/navigation_bar.0.dart' as example;
+import 'package:flutter_api_samples/material/navigation_bar/navigation_bar.1.dart' as example;
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
